Holy Moly
I have gone a little off track:

round_count counts rounds im not sure if it works yet, its initial value is 1

the looser() funciton prints a response for loosers

the winner() function prints a response for winners and first rounders

the referee function takes the player score and the highscore and returns either the looser or winner function

Im having a problem with the highsore. In the console, the code runs without an error message, but it doesnt keep track of the highscore. not sure why.